+++ core/src/main/java/apoc/nodes/Nodes.java
@@ -678,22 +678,4 @@ public class Nodes {
         return node.getDegree(relType, direction);
     }
 
-    @UserFunction("apoc.node.rebind")
-    @Description("apoc.node.rebind(node - to rebind a node (i.e. executing a Transaction.getNodeById(node.getId())  ")
-    public Node nodeRebind(@Name("node") Node node) {
-        return Util.rebind(tx, node);
-    }
-
-    @UserFunction("apoc.rel.rebind")
-    @Description("apoc.rel.rebind(rel) - to rebind a rel (i.e. executing a Transaction.getRelationshipById(rel.getId())  ")
-    public Relationship relationshipRebind(@Name("rel") Relationship rel) {
-        return Util.rebind(tx, rel);
-    }
-
-    @UserFunction("apoc.any.rebind")
-    @Description("apoc.any.rebind(Object) - to rebind any rel, node, path, map, list or combination of them (i.e. executing a Transaction.getNodeById(node.getId()) / Transaction.getRelationshipById(rel.getId()))")
-    public Object anyRebind(@Name("any") Object any) {
-        return Util.anyRebind(tx, any);
-    }
-
 }
