+++ core/src/main/java/apoc/util/Util.java
@@ -10,14 +10,12 @@ import apoc.result.VirtualRelationship;
 import org.apache.commons.io.IOUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.eclipse.collections.api.iterator.LongIterator;
-import org.neo4j.graphalgo.impl.util.PathImpl;
 import org.neo4j.graphdb.Direction;
 import org.neo4j.graphdb.Entity;
 import org.neo4j.graphdb.GraphDatabaseService;
 import org.neo4j.graphdb.Label;
 import org.neo4j.graphdb.Node;
 import org.neo4j.graphdb.NotInTransactionException;
-import org.neo4j.graphdb.Path;
 import org.neo4j.graphdb.QueryExecutionException;
 import org.neo4j.graphdb.QueryExecutionType;
 import org.neo4j.graphdb.Relationship;
@@ -25,7 +23,6 @@ import org.neo4j.graphdb.RelationshipType;
 import org.neo4j.graphdb.ResourceIterator;
 import org.neo4j.graphdb.Transaction;
 import org.neo4j.graphdb.TransactionTerminatedException;
-import org.neo4j.internal.helpers.collection.Iterables;
 import org.neo4j.internal.helpers.collection.Iterators;
 import org.neo4j.internal.helpers.collection.Pair;
 import org.neo4j.internal.kernel.api.procs.ProcedureCallContext;
@@ -851,29 +848,6 @@ public class Util {
                     }
                 })
                 .collect(Collectors.toMap(e -> e.getKey(), e -> e.getValue()));
-    }
-    
-    public static <T> T anyRebind(Transaction tx, T any) {
-        if (any instanceof Map) {
-            return (T) ((Map<String, Object>) any).entrySet().stream()
-                    .collect(Collectors.toMap(e -> e.getKey(), e -> anyRebind(tx, e.getValue())));
-        }
-        if (any instanceof Path) {
-            final Path path = (Path) any;
-            PathImpl.Builder builder = new PathImpl.Builder(Util.rebind(tx, path.startNode()));
-            for (Relationship rel: path.relationships()) {
-                builder = builder.push(Util.rebind(tx, rel));
-            }
-            return (T) builder.build();
-        }
-        if (any instanceof Iterable) {
-            return (T) Iterables.stream((Iterable) any)
-                    .map(i -> anyRebind(tx, i)).collect(Collectors.toList());
-        }
-        if (any instanceof Entity) {
-            return (T) Util.rebind(tx, (Entity) any);
-        }
-        return any;
     }
 
     public static Node rebind(Transaction tx, Node node) {
