+++ core/src/test/java/apoc/periodic/PeriodicTest.java
@@ -1,19 +1,12 @@
 package apoc.periodic;
 
-import apoc.cypher.Cypher;
-import apoc.nodes.Nodes;
 import apoc.util.MapUtil;
 import apoc.util.TestUtil;
-import apoc.util.Util;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
 import org.neo4j.common.DependencyResolver;
-import org.neo4j.configuration.GraphDatabaseSettings;
-import org.neo4j.graphdb.Label;
-import org.neo4j.graphdb.Node;
 import org.neo4j.graphdb.QueryExecutionException;
-import org.neo4j.graphdb.Relationship;
 import org.neo4j.graphdb.Result;
 import org.neo4j.graphdb.Transaction;
 import org.neo4j.graphdb.TransientTransactionFailureException;
@@ -39,7 +32,6 @@ import static apoc.periodic.Periodic.applyPlanner;
 import static apoc.util.TestUtil.testCall;
 import static apoc.util.TestUtil.testResult;
 import static apoc.util.Util.map;
-import static java.util.Collections.singletonList;
 import static java.util.stream.Collectors.toList;
 import static java.util.stream.StreamSupport.stream;
 import static org.hamcrest.CoreMatchers.equalTo;
@@ -57,12 +49,11 @@ public class PeriodicTest {
     public static final int BATCH_SIZE = 399;
 
     @Rule
-    public DbmsRule db = new ImpermanentDbmsRule()
-            .withSetting(GraphDatabaseSettings.procedure_unrestricted, singletonList("apoc.*"));
+    public DbmsRule db = new ImpermanentDbmsRule();
 
     @Before
     public void initDb() throws Exception {
-        TestUtil.registerProcedure(db, Periodic.class, Nodes.class, Cypher.class);
+        TestUtil.registerProcedure(db, Periodic.class);
         db.executeTransactionally("call apoc.periodic.list() yield name call apoc.periodic.cancel(name) yield name as name2 return count(*)");
     }
 
@@ -86,99 +77,6 @@ public class PeriodicTest {
         assertThat(count, equalTo(1L));
 
         testCall(db, callList, (r) -> assertEquals(true, r.get("done")));
-    }
-
-    @Test
-    public void testSetWithoutRebindShouldFailsOnlyWithTxSetPropertyInsteadOfSetViaCypher() {
-        // we create 2 nodes via Transaction, and we rebind only one of them
-        Node firstNode;
-        try (Transaction tx = db.beginTx()) {
-            firstNode = tx.createNode(Label.label("FirstNode"));
-            tx.commit();
-        }
-
-        try (Transaction tx = db.beginTx()) {
-            // without rebind
-            firstNode.setProperty("alpha", true);
-            fail("Should fails due to 'The transaction has been closed.'");
-            tx.commit();
-        } catch (RuntimeException e) {
-            assertTrue(e.getMessage().contains("The transaction has been closed."));
-        }
-        
-        Node secondNode;
-        try (Transaction tx = db.beginTx()) {
-            secondNode = tx.createNode(Label.label("SecondNode"));
-            tx.commit();
-        }
-
-        try (Transaction tx = db.beginTx()) {
-            // with rebind
-            secondNode = Util.rebind(tx, secondNode);
-            secondNode.setProperty("alpha", true);
-            tx.commit();
-        }
-        
-        // try periodic and runMany procs (which open a new tx) with a set statement and without rebind
-        db.executeTransactionally("CREATE (:Rebind)-[r:REL_ONE]->(:Baz)");
-
-        final Map<String, Object> runManyParams = map("cypher", "SET $n.first = 1;");
-        testCall(db, "MATCH (n:Rebind) WITH n CALL apoc.cypher.runMany($cypher, {n: n}) YIELD row RETURN row",
-                runManyParams, (row) -> {});
-        testCall(db, "MATCH (:Rebind)-[n:REL_ONE]->(:Baz) WITH n CALL apoc.cypher.runMany($cypher, {n: n}) YIELD row RETURN row",
-                runManyParams, (row) -> {});
-
-        // periodic iterate
-        final String cypherAction = "CALL apoc.cypher.runMany(\"SET $entity.test = true;\", {entity: n}) YIELD row RETURN row";
-        
-        testCall(db, "CALL apoc.periodic.iterate($cypherIterate, $cypherAction, {})", 
-                map("cypherIterate", "MATCH (n:Rebind) RETURN n", "cypherAction", cypherAction), (row) -> {});
-        testCall(db, "CALL apoc.periodic.iterate($cypherIterate, $cypherAction, {})",
-                map("cypherIterate", "MATCH (:Rebind)-[n:REL_ONE]->(:Baz) RETURN n", "cypherAction", cypherAction), (row) -> {});
-
-        testCall(db, "MATCH (n:Rebind)-[r:REL_ONE]->(:Baz) RETURN n, r", r -> {
-            final Node node = (Node) r.get("n");
-            assertEquals(true, node.getProperty("test"));
-            assertEquals(1L, node.getProperty("first"));
-            final Relationship rel = (Relationship) r.get("r");
-            assertEquals(true, rel.getProperty("test"));
-            assertEquals(1L, rel.getProperty("first"));
-        });
-    }
-    
-    @Test
-    public void testIterateRebind() throws Exception {
-
-        Node node;
-        try (Transaction tx = db.beginTx()) {
-            node = tx.createNode(Label.label("Ajeje"));
-            tx.commit();
-        }
-
-        final Map<String, Object> config = map("params", map("b", node));
-        
-        final Map<String, Object> params = map("cypherIterate", "WITH $b as b RETURN b",
-                "cypherAction", "SET b.test = 1",
-                "config", config);
-        try {
-            testCall(db, "CALL apoc.periodic.iterate($cypherIterate, $cypherAction, $config )", params, (row) -> fail());
-        } catch (Exception e) {
-            assertTrue(e.getMessage().contains("NotInTransactionException"));
-        }
-
-        // with rebind flag
-        config.put("rebind", true);
-
-        testCall(db, "CALL apoc.periodic.iterate($cypherIterate, $cypherAction, $config )", params, (row) -> {
-            assertEquals(1L, row.get("batches"));
-            assertEquals(map(), row.get("errorMessages"));
-            assertEquals(0L, row.get("failedBatches"));
-        });
-
-        testCall(db, "MATCH (b:Ajeje) RETURN b", r -> {
-            assertEquals(1L, ((Node) r.get("b")).getProperty("test"));
-        });
-
     }
 
     @Test
