--- docs/asciidoc/modules/ROOT/pages/overview/apoc.rel/apoc.rel.rebind.adoc
@@ -1,26 +1,0 @@
-////
-This file is generated by DocsTest, so don't change it!
-////
-
-= apoc.rel.rebind
-:description: This section contains reference documentation for the apoc.rel.rebind function.
-
-label:function[] label:apoc-core[]
-
-[.emphasis]
-apoc.rel.rebind(rel) - to rebind a rel (i.e. executing a Transaction.getRelationshipById(rel.getId())
-
-== Signature
-
-[source]
-----
-apoc.rel.rebind(rel :: RELATIONSHIP?) :: (RELATIONSHIP?)
-----
-
-== Input parameters
-[.procedures, opts=header]
-|===
-| Name | Type | Default 
-|rel|RELATIONSHIP?|null
-|===
-
