+++ full/src/main/resources/extended.txt
@@ -159,8 +159,11 @@ apoc.uuid.install
 apoc.uuid.list
 apoc.uuid.remove
 apoc.uuid.removeAll
+apoc.any.rebind
 apoc.coll.avgDuration
 apoc.data.email
+apoc.node.rebind
+apoc.rel.rebind
 apoc.static.get
 apoc.static.getAll
 apoc.trigger.nodesByLabel
